Created EquipmentUIController
This script manages all of the UI used for the equipment controller. It displays the difference between the current equipped equipment and incoming equipment. It controls stat sliders, affinity attribute images, and spell additions and subtractions. This system is untested and can lead to some edge cases not covered. Further testing is required along with proper implementation.