Equipment Affinity UI now works
Had to rewrite it because I assumed all of the equipment affinities would be listed, but instead only some of them are (which makes a lot of sense actually). System was rewritten into an if statement wasp nest until I simplified it down into comparing integers. Works like a charm now in my limited testing.