Created on select event for offer buttons
A new script called OfferButton was created to use the functionality of Unity Buttons while also adding an on select event for trinket selection.

An editor script was also created to make the on select event appear in the inspector while maintaining the original presentation of the button inspector functionality.