Moved the setting player prefs responsibility.
The set player pref methods have been refactor so that their only responsibility is to generated the json string rather than doing that and setting the player prefs. This was changed so that these methods can be tested to see if they return the expected results later down the line.