Moved equipment load to data manager
This had to be done since the equipment controller would not be loaded at the time time the load equipment method was called.

Equipment controller effectively acted as a middle man that didn't do anything to alter or moderate the item packet it received before saving it to the party controller anyways.