Fixed facing wall bug
This bug was quite an interesting mixture of elements. First, the position used in the set player orientation is the world position of the exit rift of the previous level. In the new level using this position, it will run the method as normal. It will then apply this final rotation to the player, then set the player back to vector zero.

Since the starting location is always at vector zero. removing the transform.position value from the equation will always run the function from the start of the new map, ensuring its success. All successful attempts before were mere coincidences.