Fixed minimap player desync bug
The bug was caused by the player controller rotating to the correct location before the combat scene was loaded. The minimap controller which handles the rotation of the minimap player icon is held within the combat controller scene. Since the player invokes the player rotation changed event when the listener isn't active, it doesn't update the minimap player icon.

I've added another event that will set the minimap player icon once the combat scene has loaded proper, eliminating this issue.