Field music now resumes from where it left off.
Added the ability to start music at any point of time, thus enabling the ability to resume music right where it left off.
The music event controller stores the current time sample from the music manager, and then forwards it back to the manager when the field music is resumed.