Volumes are now loaded when the game starts.
The reason why so many files are affected is due to the round about way this loading has to occur.

The volume is changed when the sliders are changed on the settings menu, so the game controller has to wait until the title screen is loaded, then it needs to invoke the title manager to call the set sliders function in the titleVolumeUIManager which will set the slider values, which in turn set the games volume.

There is probably a way to decouple the sliders from the mixer, but my time is limited rn, so this will have to do.