Removed continue game cheese.
The field will now regenerate to the continued players level upon continue. Before when the player pressed continue, it would allow them to explore a level 1 realm which could be easily exploited as they are far smaller than the higher level realms. Once a player enters the goal rift, the game auto saves so they could easily restart the game to climb further and further through the realms.

Players could also cheese this system by saving just before they die, but it isn't that big of a problem now that they get placed on an equally difficult floor.